# swagger.json
{
  "consumes": [
    "application.json"
  ],
  "produces": [
    "application.json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "# This is a sample recipes API",
    "title": "Recipes API",
    "contact": {
      "name": "Pavel Droo",
      "email": "123@213.com"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/recipes": {
      "get": {
        "description": "Returns list of recipes",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "listRecipes"
      },
      "post": {
        "description": "Creates new recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "newRecipe"
      }
    },
    "/recipes/search": {
      "get": {
        "description": "Search recipes by tag",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "searchRecipes",
        "parameters": [
          {
            "type": "string",
            "description": "recipe tag",
            "name": "id",
            "in": "query",
            "required": true
          }
        ]
      }
    },
    "/recipes/{:id}": {
      "get": {
        "description": "Search recipes by tag",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "searchRecipes",
        "parameters": [
          {
            "type": "string",
            "description": "ID of a recipe",
            "name": "id",
            "in": "path",
            "required": true
          }
        ]
      }
    },
    "/recipes/{id}": {
      "put": {
        "description": "Update an existing recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "updateRecipe",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the recipe",
            "name": "id",
            "in": "path",
            "required": true
          }
        ]
      },
      "delete": {
        "description": "Update an existing recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "deleteRecipe",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the recipe",
            "name": "id",
            "in": "path",
            "required": true
          }
        ]
      }
    }
  }
}